/wall

+++ source.cpp
@@ -1272,25 +1272,25 @@ set_banner:
         auto endy = std::chrono::steady_clock::now();
         const std::chrono::time_point<std::chrono::system_clock> endx = std::chrono::system_clock::now();
         auto elapsed = endy - starty;
-        unsigned long long int elapsed_time = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
+        long long int elapsed_time = std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
         const std::time_t t_start = std::chrono::system_clock::to_time_t(startx);
         const std::time_t t_end = std::chrono::system_clock::to_time_t(endx);
         std::cout << endl << S_86 << t_start << endl;
         std::cout << S_87 << t_end << endl;
-        std::cout << elapsed_time * 1.0 / 1000000 << S_84 << endl;
+        std::cout << static_cast<double>(elapsed_time) * 1.0 / 1000000 << S_84 << endl;
         if (five_count == 0) {
             std::cout << endl << S_73 << countx << S_74 << endl
-                << S_75 << five_count << "  " << five_count * 100.0 / countx << "%" << endl
-                << S_76 << four_count << "  " << four_count * 100.0 / countx << "%" << endl
+                << S_75 << five_count << "  " << static_cast<double>(five_count) * 100.0 / static_cast<double>(countx) << "%" << endl
+                << S_76 << four_count << "  " << static_cast<double>(four_count) * 100.0 / static_cast<double>(countx) << "%" << endl
                 << S_77 << five_count_c << " : " << five_count_w << " : " << four_count_c << " : " << four_count_w << endl << endl;
         }
         else if (five_count > 0) {
             std::cout << endl << S_73 << countx << S_74 << endl
-                << S_75 << five_count << "  " << five_count * 100.0 / countx << "%" << endl
-                << S_76 << four_count << "  " << four_count * 100.0 / countx << "%" << endl
+                << S_75 << five_count << "  " << static_cast<double>(five_count) * 100.0 / static_cast<double>(countx) << "%" << endl
+                << S_76 << four_count << "  " << static_cast<double>(four_count) * 100.0 / static_cast<double>(countx) << "%" << endl
                 << S_78 << max_fives << S_79 << max_fivesth << S_80 << max_fivecount + 1 << S_81 << endl
                 << S_82 << min_fives << S_79 << min_fivesth << S_80 << min_fivecount + 1 << S_81 << endl
-                << S_83 << ave_fives * 1.0 / five_count << endl
+                << S_83 << static_cast<double>(ave_fives) * 1.0 / static_cast<double>(five_count) << endl
                 << S_77 << five_count_c << " : " << five_count_w << " : " << four_count_c << " : " << four_count_w << endl << endl;
         }
         else { std::cout << E_8 << endl; goto full_quit; }
@@ -1312,17 +1312,17 @@ set_banner:
         if (chosen_banner == 5 || four_count == 0 || five_count == 0) { wishes_number = 0; std::cout << S_72 << endl; goto print_2; }
         else if (chosen_banner == 1 || chosen_banner == 2 || chosen_banner == 4) {
             std::cout << S_88 << endl << endl;
-            for (int iout = 0; iout < 10; iout++) { std::cout << iout + 1 << "  " << four_pity[iout] << "   " << four_pity[iout] * 100.0 / four_count << "%" << endl; }
-            std::cout << "10+  " << four_pity[10] << "   " << four_pity[10] * 100.0 / four_count << "%" << endl << endl << S_89 << endl << endl;
-            for (int iout = 0; iout < 90; iout++) { std::cout << iout + 1 << "  " << five_pity[iout] << "   " << five_pity[iout] * 100.0 / five_count << "%" << endl; }
+            for (int iout = 0; iout < 10; iout++) { std::cout << iout + 1 << "  " << four_pity[iout] << "   " << static_cast<double>(four_pity[iout]) * 100.0 / static_cast<double>(four_count) << "%" << endl; }
+            std::cout << "10+  " << four_pity[10] << "   " << static_cast<double>(four_pity[10]) * 100.0 / static_cast<double>(four_count) << "%" << endl << endl << S_89 << endl << endl;
+            for (int iout = 0; iout < 90; iout++) { std::cout << iout + 1 << "  " << five_pity[iout] << "   " << static_cast<double>(five_pity[iout]) * 100.0 / static_cast<double>(five_count) << "%" << endl; }
             wishes_number = 0;
             goto print_2;
         }
         else if (chosen_banner == 3) {
             std::cout << S_88 << endl << endl;
-            for (int iout = 0; iout < 10; iout++) { std::cout << iout + 1 << "  " << four_pity[iout] << "   " << four_pity[iout] * 100.0 / four_count << "%" << endl; }
-            std::cout << "10+  " << four_pity[10] << "   " << four_pity[10] * 100.0 / four_count << "%" << endl << endl << S_89 << endl << endl;
-            for (int iout = 0; iout < 80; iout++) { std::cout << iout + 1 << "  " << five_pity_w[iout] << "   " << five_pity_w[iout] * 100.0 / five_count << "%" << endl; }
+            for (int iout = 0; iout < 10; iout++) { std::cout << iout + 1 << "  " << four_pity[iout] << "   " << static_cast<double>(four_pity[iout]) * 100.0 / static_cast<double>(four_count) << "%" << endl; }
+            std::cout << "10+  " << four_pity[10] << "   " << static_cast<double>(four_pity[10]) * 100.0 / static_cast<double>(four_count) << "%" << endl << endl << S_89 << endl << endl;
+            for (int iout = 0; iout < 80; iout++) { std::cout << iout + 1 << "  " << five_pity_w[iout] << "   " << static_cast<double>(five_pity_w[iout]) * 100.0 / static_cast<double>(five_count) << "%" << endl; }
             wishes_number = 0;
             goto print_2;
         }
